app/requirements.txt
pydantic
fastapi
fastapi[all]
uvicorn
sqlalchemy
apscheduler
python-jose
python-jose[cryptography]
passlib
python-multipart
bcrypt
python-dotenv